Sec 6 - 46. Nav Link v/s anchor tag
NavLink is intelligent enough to remember the route and highlight the menu item to be highlighted, while <a> tag is not.

in the example which is changed in this file, Bind props doesnt get highlighted as we use the <a> tag while the others get highlighed on click as they use the NavLink.

https://slk.udemy.com/course/introduction-to-aspnet-core-x/learn/lecture/45108315#overview